requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiosignal==1.3.1
attrs==24.2.0
blinker==1.9.0
click==8.1.7
Flask==2.0.1
Frozen-Flask==1.0.2
frozenlist==1.5.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
multidict==6.1.0
packaging==24.2
propcache==0.2.0
python-dotenv==1.0.1
tesla-api==2.0.1
Werkzeug==2.0.3
yarl==1.18.0
requests==2.31.0
asgiref==3.7.2
teslapy==2.8.0